custom-elements/instantiating-custom-elements/custom-element-type-is-attribute-test-001.html

<!DOCTYPE html>
<html>
<head>
<title>Custom element type is  the value of the is attribute of the custom element</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="The custom element type can be given to a custom element at the time of its instantation as  the value of the is attribute of the custom element">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#instantiating-custom-elements">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='../testcommon.js'></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

function doTest(element) {
    var doc = newHTMLDocument();
    
    var name = 'x-foo-' + element;
    
    var XFoo = doc.register(name);
    
    if (HTML5_TABLE_ELEMENTS.indexOf(element) != -1) {
        doc.body.innerHTML = 
            '<table>' +
            '<' + element + ' id="customElement" is="' + name + '"></' + element + '>' +
            '</table>';
    } else {
        doc.body.innerHTML = 
            '<' + element + ' id="customElement" is="' + name + '"></' + element + '>';        
    }
    
    var xfoo = doc.querySelector('#customElement');

    var obj = doc.createElement(element);
    
    assert_class_string(xfoo, obj.constructor.name,
            'Custom element  must be a ' + obj.constructor.name);
}


var parameters = [];

for (var i = 0; i < HTML5_ELEMENTS.length; i++) {
    if (HTML5_DOCUMENT_ELEMENTS.indexOf(HTML5_ELEMENTS[i]) == -1) {
        parameters[i] = ['Custom element type ' + HTML5_ELEMENTS[i], HTML5_ELEMENTS[i]];
    }
}

generate_tests(doTest, parameters, 
        'Custom element type is the local name of the custom element. '
        + 'Element is created in HTML');

</script>
</body>
</html>
